portfolio-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
requests==2.32.3
requests-toolbelt==1.0.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
opencv-python==4.9.0.80
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
dino @ git+https://github.com/facebookresearch/dino.git 